requirements.txt
pillow == 10.2.0
flask == 3.0.2
pip==24.0
gunicorn==21.2.0
reportlab~=4.1.0